# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FundsLink Academy API
  version: "1.1.0"
  description: >
    Contract-first API (S2.7) for FundsLink v1 — locked WITH ERD-PACKAGE v1.0.
    Governed by the master-spec, technical-architecture, and database doctrine.
    FastAPI's generated schema is CI-diffed against this file; drift fails the build.
  contact: { name: "Maluleke Kurhula Success — Founder, KSDRILL SA" }
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      operationId: authRegister
      security: []
      summary: Register a new student account (consent recorded — MASTER-SPEC §15.4)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
  /auth/login:
    post:
      operationId: authLogin
      security: []
      summary: Login (rate-limited; lockout per TAD §3.1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      operationId: authRefresh
      security: []
      summary: Rotate refresh token (family revocation on reuse)
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/mfa/enroll:
    post:
      operationId: authMfaEnroll
      summary: Begin TOTP MFA enrolment — returns provisioning URI + recovery codes (shown once)
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MfaEnrollResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/mfa/activate:
    post:
      operationId: authMfaActivate
      summary: Activate MFA by confirming a TOTP code (mandatory for privileged roles — TAD §3.1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MfaActivateRequest' }
      responses:
        '204': { description: MFA activated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Unauthorized: { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Conflict:     { description: Business-rule conflict (stable error.code), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Validation:   { description: Validation failure, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited:  { description: Rate limited (Retry-After set), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
    RegisterRequest:
      type: object
      required: [email, password, consents]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 10 }
        consents:
          type: array
          items:
            type: object
            required: [purpose, wording_version]
            properties:
              purpose: { type: string }
              wording_version: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        mfa_code: { type: string, description: Required for privileged roles }
    AuthTokens:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: bearer }
        expires_in: { type: integer }
    MfaEnrollResponse:
      type: object
      required: [secret, provisioning_uri, recovery_codes]
      properties:
        secret: { type: string, description: Base32 TOTP secret (shown once) }
        provisioning_uri: { type: string, description: otpauth:// URI for the QR code }
        recovery_codes: { type: array, items: { type: string }, description: Single-use backup codes (shown once) }
    MfaActivateRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, description: A current 6-digit TOTP from the authenticator }
...